+++ crates/rils_execution/src/value/display.rs
@@ -35,6 +35,7 @@ impl fmt::Display for Value {
             Self::Rc(_) => write!(f, "<Rc>"),
             Self::Weak(_) => write!(f, "<Weak>"),
             Self::Cell(_) => write!(f, "<Cell>"),
+            Self::RefCell(_) => write!(f, "<RefCell>"),
             Self::SequenceIterator(_) => write!(f, "<sequence iterator>"),
             Self::BytecodeIterator(_) => write!(f, "<bytecode iterator>"),
             Self::Reference(reference) => match reference.read() {